requirement.txt
streamlit
PyMuPDF
pillow
pytesseract
langchain
langchain-community
langchain-ollama
langchain-chroma
sentence-transformers
chromadb
ollama
